Good stopping point to deploy
