+++ client/src/components/DetailDialog/DetailDialog.jsx
@@ -1,0 +1,34 @@
+import React from "react";
+import TextField from "@material-ui/core/TextField";
+import Dialog from "@material-ui/core/Dialog";
+import DialogActions from "@material-ui/core/DialogActions";
+import DialogContent from "@material-ui/core/DialogContent";
+import DialogContentText from "@material-ui/core/DialogContentText";
+import DialogTitle from "@material-ui/core/DialogTitle";
+import MenuItem from "@material-ui/core/MenuItem";
+
+export default function DetailDialog(props) {
+    const [open, setOpen] = useState(false);
+
+    const handleClickOpen = () => {
+		setOpen(true);
+	};
+
+	const handleClose = () => {
+		setOpen(false);
+    };
+
+	return (
+		<div>
+			<Dialog
+				open={open}
+				onClose={handleClose}
+				aria-labelledby="form-dialog-title"
+			>
+				<DialogContent>
+					<DialogContentText>{wine.name}</DialogContentText>
+				</DialogContent>
+			</Dialog>
+		</div>
+	);
+}
