+++ client/src/screens/Champagnes/Champagnes.jsx
@@ -1,11 +1,20 @@
-import React from 'react'
+import React from "react";
+import { Link } from "react-router-dom";
 
 export default function Champagnes(props) {
-    const { currentUser, wines } = props;
-    
-    return (
-        <div>
-            Poppin' bottles biyyyyaaatttchh..
-        </div>
-    )
+	const { wines } = props;
+
+	return (
+		<div className="white_wines">
+			{wines
+				.filter((wine) => {
+					return wine.category === "Champagne";
+				})
+				.map((wine) => (
+					<Link className="card" to={`/wine/reds`}>
+						<div>{wine.name}</div>
+					</Link>
+				))}
+		</div>
+	);
 }
