+++ client/src/screens/WineDetail/WineDetail.css
@@ -1,0 +1,12 @@
+.wine_details {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: flex-start;
+    padding:0;
+}
+
+.wine_details h4 {
+    margin: 5px;
+    padding: 0;
+}
