+++ client/src/screens/Wines/Wines.css
@@ -1,6 +1,6 @@
 .wine_page{
-    width: 90vw;
-    margin: 0 5vw 0 5vw;
+    width: 80vw;
+    margin: 0 10vw 0 10vw;
 }
 
 .wines_table_header {
@@ -33,14 +33,3 @@
     grid-column-start: 4;
 }
 
-.min_price {
-    grid-column-start: 5;
-}
-
-.max_price {
-    grid-column-start: 6;
-}
-
-.size {
-    grid-column-start: 7;
-}
